# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Domentation of supermenu API",
        "version": "1.0.0",
        "title": "supermenu APIs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "davidshumbusho10@gmail.com",
            "name": "David Shumbusho"
        },
        "licence": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
...
    "paths": {
...
        "/resto/search/{name}": {
            "get":{
            "tags": [
                "Resto"
            ],
            "summary": "search restaurant by name",
            "description": "search restaurant by name",
            "operationId": "search restaurant by name",
            "consumes": [
                "application/json"
            ],
            "produces": [
                "application/json"
            ],
            "parameters": [
                {
                    "in": "path",
                    "name": "name",
                    "required": true,
                    "type": "string"
                }
            ],
            "responses": {
                "200": {
                    "description": "successful operation"
                },
                "400": {
                    "description": "Invalid input"
                }
            }
        }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}